+++ CHANGELOG.md
@@ -20,13 +20,16 @@ changes from that point onwards.
 * Management SDK now properly raises errors and works properly with localized content
 * Queries now properly returns `::Contentful::Array` instead of `::Array` instances [#67](https://github.com/contentful/contentful_model/issues/67)
 * Queries now allow multiple fields for `find_by` and also works with `sys` fields [#49](https://github.com/contentful/contentful_model/issues/49)
+* Queries now do not share state any more, interface remained the same [#94](https://github.com/contentful/contentful_model/issues/94)
+* Queries using `find_by` can now also request `sys` fields without having to specify always the `sys.` prefix
 
 ### Changed
 
 * Updated to CDA SDK 2.x
 * Updated Specs to run mostly against real data instead of mocks
 * Validations are now run by default before `#save`
-* Removed `require_all` dependency
+* Removed `require_all` dependency [#95](https://github.com/contentful/contentful_model/issues/95)
+* `ChainableQueries` module removed, behaviour now moved to `Queries` and `Query`. [#94](https://github.com/contentful/contentful_model/issues/94)
 
 ## 0.2.0